thrivetracker-backend/requirements.txt
asgiref==3.6.0
Babel==2.12.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.52.0
django-cors-headers==3.14.0
django-money==3.0.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
psycopg2-binary==2.9.6
py-moneyed==2.0
pycparser==2.21
PyJWT==1.7.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
